Fix 2PC file check logic in PrepareRedoAdd()
In commit cb0cca1880723b4c90c56cdcf3842489ef036800,
the 2PC state file check logic is added in PrepareRedoAdd()
to fix the bug of 2PC recovery. Each time a file is accessed,
the OS creates a dentry for it and increases the reference
count of the parent directory's dentry. The type of the dentry
reference count is int32. When there are many 2PC transactions,
this logic may lead to an overflow of the parent dentry's
reference count, which could ultimately cause the OS to crash.
Since the 2PC data on disk is only restored at the beginning
of recovery with restoreTwoPhaseData(), after reaching a
consistent state, the 2PC transactions are only added from the
WAL. So we perform the file access check only when the server
has not yet reached a consistent state. Once consistency has
been reached, if a duplicate 2PC transaction is added, it will
directly result in an error in the subsequent replay logic.