Document AccessExclusive lock behaviour on standbys
Access exclusive lock is the only lock propagated through the WAL and
replayed on standbys. This can lead to confusing situations as some
users assume the lock is restricted on the primary and don't expect the
locked table to be unreadable on the standbys.

This patch adds more details on the expected behaviour of Access
Exclusive lock in the table-level locks documentation.