PoC: introduce CustomJoin as separate from CustomScan
Introduce a new CustomJoin node, intended for out-of-core join nodes.
The CustomScan is supposed to allow that, but there's a couple pieces
that make it inconvenient. The targetlist uses INDEX_VAR, so it's not
clear which relation the Vars come from, etc.

So this adds a new node, descended from Join (not Scan), with all the
processing specific to joins (in setrefs etc.).

Experimental. Does are not updated, some bits may be irrelevant for
joins, etc.