Give StartReadBuffers() a more robust interface.
Remove the need for read_stream.c to initialize and clear
stream->buffers[] elements that will be passed to StartReadBuffers(),
and to count them on return.  Instead, add an explicit npinned argument
to tell StartReadBuffers() how many buffers were already pinned by the
previous call (if it was a short read), and to learn how many are pinned
beyond *nblocks after this call returns (if it is a short read).  It is
an in/out argument.  The output of one call can be passed directly as
input to the next call, just like the buffers themselves.

The new function argument count still fits in registers in Unixoid
calling conventions.  It doesn't on Windows/x86, but it didn't before
either.  The single-buffer StartReadBuffer() specialization is not
affected.

read_stream.c now uses stream->pending_read_npinned to carry that number
from call to call, replacing stream->forwarded_buffers.  It no longer
maintains it itself, which now seems a little fragile, layer-violating
and expensive.  It still needs to consult the value when respecting pin
limits, but other than that it's entirely StartReadBuffers()' job to
worry about it.  The new member variable name better reflects its
relationship with pending_read_{blocknum,nblocks}.  It is the number of
buffers in that range that were already pinned by an earlier short read.
After a short read, _blocknum advances, _nblocks shrinks, and _npinned
has already been filled in by StartReadBuffers(), so we're ready for the
next call at next_buffer_index, where *npinned buffers have been placed
by StartReadBuffers().

Assertions about the identity of forwarded buffers are removed, because
StartReadBuffers() already has such assertions, and assertions about
queue entries after that are no longer relevant.

Reviewed-by: Xuneng Zhou <xunengzhou@gmail.com>
Discussion: https://postgr.es/m/19006-80fcaaf69000377e%40postgresql.org