Change gb-18030-2000.xml to gb-18030-2000.ucm
This is the first in a series of patches to upgrade GB18030 encoding
from the 2000 standard to the 2022 standard.

In this patch, gb-18030-2000.xml is replaced with gb-18030-2000.ucm,
and UCS_to_GB18030.pl is updated to process .ucm files.

Note that the gb-18030-2000.ucm used here is not the latest version
from GitHub. A newer version exists that fixes a mapping error. To
keep this patch focused, we are using a version of gb-18030-2000.ucm
that matches gb-18030-2000.xml, so the mapping files remain unchanged
in this commit.

The next patch will update gb-18030-2000.ucm to the latest version.

Author: Chao Li <lic@highgo.com>