Pre-check potential XLogRecord oversize.
XLogRecord size check is placed in critical section and in case
of failure PANIC will be generated. It seems not good, so to
avoid this, we can calculate approximate xlog record size before
critical section and check it.

Size prediction is based on xlog update and xlog delete logic and can
be revised in case of it changing, now buf size is limited by
UINT16_MAX(Assert(regbuf->rdata_len <= UINT16_MAX) in xloginsert).

Anyway to accommodate some overhead, 1M is substract from predicted
value. It seems now it is quite enough.