jsonb: Optimize JsonbContainerTypeName by reordering type checks
The JsonbContainerTypeName() function currently checks for the less
common scalar container type before checking for objects and arrays.

This commit reorders the checks to prioritize the most common cases. The
macros JsonContainerIsArray() and JsonContainerIsObject() are
simple bit checks and are now evaluated first. This avoids the
overhead of calling the JsonbExtractScalar() function in the vast
majority of use cases.

Author: Chao Li <lic@highgo.com>