Fix pg_waldump to exit cleanly at end of WAL
Previously, running pg_waldump without the -e option would always
produce an error message when reaching the end of the WAL, because
the parser encountered a zero-length record. This patch changes
the behavior so that pg_waldump exits gracefully when the last
valid WAL record has been processed, without printing an unnecessary
error.

Author: Chao Li <lic@highgo.com>