Composite data types have been added to the GUC
The new module of the configuration system allows you to map the values of
composite data types into composite variables (structures, static arrays,
dynamic arrays) in the C code in a unified way.

Mapping works due to the signature that the programmer declares for the
composite type.  For core options, the declaration data is in the
UserDefinedConfigureTypes array. Composite types for options from
extensions are declared using the DefineCustomCompositeType function.
All declarations must be arranged topologically. That is, if the type A
option contains a type B field, then type B must first be declared,
and then type A.

The main fields in the type definition are the type name and its signature.
The type signature has the following syntax:
“field_type field_name; field_type field_name; ...; field_type field_name”
where field_type is the already registered type, field_name is the field name.

There are also data types that do not need to be declared - these are arrays.
So, if there is a registered type A, then data types automatically become
available: A[n] is a static array of length n and A[] is a dynamic array.

Note that the declared type signature must match the signature of the structure
in the C code, since it will then be used to calculate the alignment of fields
according to the rules of the C language.
Dynamic arrays are always mapped into a structure like:
struct DynArr {
	void *data; //pointer to data
	int size; //length of the array
}

After declaring the type, you can declare a composite type configuration
option. The core options are declared in the guc_parameters.dat file. They must
specify the type => ‘composite’ fields and specify in the type_name field
the name of the composite type that was declared earlier. In the boot_val
field, write a pointer to a global variable that will store this value.
Options from extensions are declared using DefineCustomCompositeVariable.

Now you can use the following syntax to work with the new options both in
the configuration file and in psql:

Field access: option_name->field_name
Access to an array element: option_name[index]
You can combine these access methods.
Dynamic arrays always have implicit fields data and size.
data is the data of the array, size is its length.

Values of composite types have the following syntax:
- Structures: {field: value, ..., field: value}
- Static arrays: [index: value, index: value]
- Dynamic arrays arrays have implicit fields, so you can use 2 syntaxes:
-- compact (same as for static arrays) and
-- extended:
   {data: [index: value, .., index: value], size: value}.
It is not necessary to write indexes in array values. If you write
without indexes, it is assumed that indexing starts from 0 with an increment
of 1. In this case, all elements within the same array must be either with
or without indexes.

When using the show command, the display of the dynamic array depends on the
extended_guc_arrays option. If this flag is true, then the extended form
is used, otherwise the compact form is used.

String values within composite types also support escape sequences.
All the functionality available to scalar options is also

The system uses incremental semantics. This means that when writing to a .conf
file or the set command, only the specified fields of the structure will be
changed, the remaining fields will not be involved. The same applies to the
alter system see. When using this command, the current value will be written
to the .auto.conf file with the changed fields that were described when
calling the command, while the current value of the option will not change.