Refactor external TOAST pointer code for better pluggability
This commit introduces a new interface for external TOAST pointers,
which is able to make a translation of the varlena pointers stored on
disk to/from an new in-memory structure called toast_external.  The
types of varatt_external supported on disk need to be registered into a
new subsystem in a new file, called toast_external.[c|h], then define a
set of callbacks to allow the toasting and detoasting code to use it.

A follow-up change will rely on this refactoring to introduce new
vartag_external values with an associated varatt_external_* that is
able, which would be used in int8 TOAST tables.