WIP: pg_service: implement defaults section
Default connection options may now be overridden by creating a section
at the top of a pg_service.conf file with the following header:

  [any-arbitrary-section-name]
  +=defaults
  host=...

System-level defaults (from the service file residing in PGSYSCONFDIR)
and user-level defaults (from the PGSERVICEFILE) are merged, with the
user-level options taking precedence.

With this change, connection options fall back in order of decreasing
precedence:

1) Connection string
2) User service
  2b) System service, but only if no user service was found
3) Environment variable
4) User default
5) System default
6) Compile-time default

Some code complexity here is due to the fact that the service name in
use may depend on the defaults. The new tests make use of this to ensure
that Test::Cluster environment variables are overridden as needed.