Minor fix to avoid maybe-uninitialized warning related to notnull_relpos
This moves the initialization of notnull_relpos to after the block that
the seems to be overlooked by the compiler and just before the value is
used in the do/while test.  The code appears safe as is, this change
removes the ambiguity that confuses the compiler triggering the "maybe
unitialized" warning.