Make AIO handle callbacks extensible
With the extensible SMGR patch, extensions can register their own
storage manager logic. However, this is not that helpful with the startreadv
as the AIO API provides no extensibility mechanisms:

* There's no way to extend the callback chain defined by md.c, as it is
  not visible to extensions
* And there's also no way to define another completely custom chain
  instead of it

This patch aims to improve this by making the md callback chain visible
to extensions, and also increasing the maximum number of callbacks so
that a few extensions can safely register their additions to it.