WIP: pg_service: implement PGNODEFAULTS
Tests need to ensure that they have full control of the default
connection state. While PGSYSCONFDIR was already redirected as needed,
the user's config file would still be consulted when PGSERVICEFILE was
empty. Add a PGNODEFAULTS envvar to turn off the previous feature, and
make use of it in pg_regress and Test::Cluster connections.