Allow UTF-8 locales to use strcoll_l() on Windows.
On Windows we allow locales with non-UTF-8 encodings in UTF-8 databases
for historical reasons, and convert to wchar_t when collating strings.
Allow plain strcoll_l() to be reached instead of wcscoll_l() when the
locale is a UTF-8.

XXX Does this work as expected?

XXX How is the performance?  It might be converting to wchar_t
internally anyway, depending on whether it can work incrementally like
ICU.