initdb: if locale is C or C.UTF-8, use builtin provider.
If the provider is unspecified, use the builtin provider C or
C.UTF-8. If the provider is specified, then do not override it.

The C locale has always been, effectively, the builtin provider, in
the sense that it uses built-in logic rather than strcoll(), etc. The
change here is mostly about the catalog representation.

The C.UTF-8 locale has used libc, but by doing so, collation doesn't
benefit from important performance optimizations. Now that we have a
builtin "C.UTF-8" collation which does benefit from those
optimizations, use that.