Reorganize heap update logic and update simple_heap_update()
NOTE:
Taken from the cf-5556 work to demonstrate how the two fit together.
This work addresses the catalog tuple update path, the other CF is
about updates due to executor or replication.

This commit refactors the interaction between heap_tuple_update(),
heap_update(), and simple_heap_update() to improve code organization and
flexibility. The changes are functionally equivalent to the previous
implementation except that now updates to catalog tuples no longer
invoke HeapDetermineColumnsInfo() to determine which indexed attributes changed.
The set of modified attributes is now passed into simple_heap_update()
which then intersects that with the indexed attributes.  The remaining
logic from HeapDetermineColumnsInfo() related to replica identity still
exists in simple_heap_update().

As part of this reorganization, the handling of replica identity key
attributes has been adjusted. Instead of fetching a second copy of
the bitmap during an update operation, the caller is now required to
provide it. This change applies to both heap_update() and
heap_delete().