Allow unique GiST indexes
- Added tests showing you can CREATE UNIQUE INDEX ... USING gist, with correct
  semantics building the index and adding to the index.
- Updated the gist README.
- Updated documentation.
- Added excludeFn to the GISTSTATE, used to compare values.
- Added gist_check_unique to do the work. We call this once we have found the
  leaf page, then we check all tuples on the page. As an optimization we could
  first check the consistent function.

Because GiST is lossy (at least if the opclass provides a compress function),
the Datum in the index might not match the Datum in the heap. But the excludeFn
needs heap Datums. So we have to load them before checking. That is not really a
big performance hit though, because we need the heap tuples anyway for
visibility information.

If we detect that we are backing a WITHOUT OVERLAPS constraint (because
indisexclusion is set at the same time as indisunique), then we use overlaps for
the last element's excludeFn. Also we adjust our error messages to match what
the exclusion constraint would produce. This means that almost all the work in
check_exclusion_or_unique_constraint is redundant. Perhaps we should detect that
for WITHOUT OVERLAPS constraints we can rely on the index to have checked that
already? (It still needs to reject empty values though.) In a couple cases I had
to update tests for CREATE CONSTRAINT ... WITHOUT OVERLAPS, because we now
encounter a different conflicting tuple.

I haven't dealt with MVCC issues yet. When we find a not-yet-committed
conflicting tuple, we need to wait and re-check after that transaction has
finished. See nbtree for how this needs to work.

Also I'm trying to handle NULLS NOT DISTINCT correctly. In nbtree the code
assumes that nulls are never equal. So is that handled outside of specific IAMs?