Resync logical replication slot after switchover
When the former primary is started up after clean shutdown,
it was refusing to synchronize the logical failover slots with the new primary
because synced flag is false.
The following error message is produced: "exiting from slot synchronization because same
slot name already exists on the standby".

If local slot has failover=true, it can be overwrite to be synced with
the primary.

This patch mainly change the drop_local_obsolete_slots() function and
check that a local failover slot could be dropped.