Mention that md5-hashed passwords are deprecated in sample postgresql.conf.
Version 18 deprecates passwords hashed with password_encryption = 'md5', but
the comments for this GUC in postgresql.conf.sample did not mention this.
Adding a deprecation notice here might make more people aware of this and
lead them to migrate to SCRAM. While at, add a comment to the
md5_password_warnings GUC mentioning the MD5 deprecation there as well.