Update the remainder of catalog updates using the new APIs
NOTE: This is a work-in-progress commit, some of the files (aclchk.c,
heap.c) have been adapted to use the proposed macros, the rest remain
unchanged and use the ealier macros.

Apply all necessary changes to use the new macros for catalog tuples.
This commit finishes the work started in the previous one.  There should
be no behavioral changes resulting from these commits.