Avoid DISTINCT step if the input path already generates distinct rows.
This patch teaches the planner to recognize if the output of the scan/join
path happens to satisfy the DISTINCT clause of the query. If it does, paths
like UniquePath or AggPath do not have to be applied to it.

The problem can be addressed by adding uniqueness information ("unique keys")
to the table / index scan paths and propagate it to the upper paths
(bottom-up), like we do with ordering information ("path keys"). However it's
not quite clear how those "unique keys" should participate in planning
decisions, which are already non-trivial. Therefore it doesn't seem worth
adding extra information to Path and subclasses.

This patch handles the problem the other way round (top-down): the scan/join
path at the top of the tree is checked, and if it has subpaths, the check
includes recursion into these.

* Index path is considered distinct if the index is unique and if the index
  keys are a subset of the DISTINCT expressions referencing the index's parent
  table. No DISTINCT expression may reference a nullable column.

* Sequential scan path is considered distinct if the underlying table has at
  least one index for which a distinct index path can be created (see
  above). No DISTINCT expression may reference a nullable column.

* Unique path is considered distinct if its keys are a subset of the DISTINCT
  expressions referencing the input path.

* Join path is considered distinct if 1) the expressions belonging to the
  outer path form unique rows and 2) there is no more than one inner row per
  output row.

* Paths that do not affect uniqueness are considered distinct if their input
  paths are so.

The existing planner infrastructure is used, i.e. no fields are added to Path
and subclasses.