Change default_toast_compression to lz4
The default value for default_toast_compression was pglz. The main
reason is that this option is always available. However, it is known
that pglz uses more CPU than lz4. The default value will be lz4 if lz4
support is built, otherwise, pglz.

The autoconf behavior is to provide an error if lz4 dependency is not
found. You should inform --witthout-lz4 if lz4 is not available. The
meson behavior is to try to detect lz4, if not found, defaults to pglz.
It acts like ICU or readline that you have to opt out if you don't want
lz4 support.

While at it, remove quotes from default_toast_compression. We usually do
not use quotes for enum.