add in-memory btree tuple index
This patch implements in-memory B+tree structure. It will be used as
index for special type of grouping using index.

Size of each node is set using macro. For convenience equals 2^n - 1, so
for internal nodes we effectively calculate size of each page and find
split node (exactly in the middle), and for leaf nodes we can distribute
tuples for each node uniformely (according to the newly inserted tuple).

It supports different memory contexts for tracking memory allocations.
And just like in TupleHashTable during Lookup it uses 'isnew' pointer to
prevent new tuple creation (i.e. when memory limit is reached).

Also it has key abbreviation optimization support like in tuplesort. But
some code was copied and looks exactly the same way, so it is worth
separating such logic into a separate function.