Add _pq_.cursor protocol extension for cursor options
Add a protocol extension that allows clients to pass cursor option
flags in Bind messages, enabling HOLD, SCROLL, and NO_SCROLL on
named portals.

The extension appends the options to an optional Int32
field to the Bind message when negotiated during connection
startup.

The cursor_protocol connection parameter controls if the
extension is enabled.

Add PQsendBindWithCursorOptions() to libpq, which sends
Bind+Describe to create a named portal with the
cursor options.

Also, a new test module is added.