Don't use deprecated and insecure PQcancel psql and other tools anymore
All of our frontend tools that used our fe_utils to cancel queries,
including psql, still used PQcancel to send cancel requests to the
server. That function is insecure, because it does not use encryption to
send the cancel request. This starts using the new cancellation APIs
(introduced in 61461a300) for all these frontend tools. These APIs use
the same encryption settings as the connection that's being cancelled.
Since these APIs are not signal-safe this required a refactor to not
send the cancel request in a signal handler anymore, but instead using a
dedicated thread. Similar logic was already used for Windows anyway, so
this also has the benefit that it makes the cancel logic more uniform
across our supported platforms.

The calls to PQcancel in pg_dump are still kept and will be removed in
a later commit. The reason for that is that that code does not use
the helpers from fe_utils to cancel queries, and instead implements its
own logic.