Fix unsupported relations for publications
Although publication check for unsupported relations for the table list,
it is not checking the partitions of a partitioned table. The
consequence is that data from this partitioned table is not replicated.

Inheritance is already covered because parent and children tables are
included in the publication-relation mapping. There is no test coverage
for this case so add it.

This is not a complete solution. SQL commands that manipulate the
partitioned table (ALTER TABLE ... ATTACH PARTITION, ALTER TABLE ... SET
UNLOGGED, CREATE TABLE ...  PARTITION OF) might break this contract.