TableAM: Support AM-specific fast visibility tests
Previously, we assumed VM_ALL_VISIBLE(...) is universal across all
AMs. This is probably not the case, so we introduce a new table
method called "table_index_vischeck_tuples" which allows anyone to
ask the AM whether a tuple (or list of tuples) is definitely visible
to us, or might be deleted or otherwise invisible.

We implement that method directly for HeapAM; usage of the facility
will follow in later commits.