IOS: Support tableAM-powered prechecked visibility statuses
Previously, we assumed VM_ALL_VISIBLE(...) is universal across all
AMs.  Now, we use the new table_index_vischeck_tuples API, and allow
index AMs to do this check as well.  This removes the need for index
AMs to hold on to pinned pages in index-only scans, as long as they
do the visibility checks on the tuples they pull from the pages
whilst they still have a pin on the page.

selfuncs.c's get_actual_variable_endpoint() is similarly updated
to use this new visibility checking infrastructure.

Future commits will implement this new infrastructure in gist,
sp-gist, and nbtree.