Introduce ENDLIST to terminate multiline makefile lists
Introduce a common ENDLIST variable and append it to these multiline lists,
so that items can be added or removed without worrying about whether they are
last in the list. This is purely a build-system cleanup and does not change
the resulting list contents.

Define ENDLIST once in Makefile.global.in and apply it consistently across
the affected makefiles.

Author: Chao Li <lic@highgo.com>